% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataref.R
\docType{data}
\name{areaciem}
\alias{areaciem}
\title{areaciem}
\format{
An object of class \code{data.frame} with 6501 rows and 5 columns.
}
\usage{
areaciem
}
\description{
Spatial definition of ICES stocks (2024 version)
}
\keyword{datasets}
