% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0001_ceData-class.R
\docType{class}
\name{ceData-class}
\alias{ceData-class}
\title{ceData-class}
\description{
Fishing effort object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{ce}}{Dataframe of 19 columns}
}}

