% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0001_ceData-class.R
\docType{methods}
\name{ceData}
\alias{ceData}
\alias{ceData,data.frame-method}
\alias{ceData,missing-method}
\title{ceData constructors}
\usage{
ceData(ce, ...)

\S4method{ceData}{data.frame}(ce, desc = "Unknown stock", ...)

\S4method{ceData}{missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{ce}{dataframe following the ce slot of a \link{ceData-class} object}

\item{desc}{descriptor}

\item{missing}{: create an empty \link{ceData-class} object}

\item{\\dots}{parameters}
}
\description{
ceData constructors
}
