% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0001_ceData-class.R
\docType{methods}
\name{ce}
\alias{ce}
\alias{ce,ceData-method}
\alias{head,ceData-method}
\alias{tail,ceData-method}
\alias{summary,ceData-method}
\alias{dim,ceData-method}
\alias{rbind2,ceData,ceData-method}
\alias{subset,ceData-method}
\title{ceData-methods}
\usage{
ce(object, ...)

\S4method{ce}{ceData}(object, ...)

\S4method{head}{ceData}(x, ...)

\S4method{tail}{ceData}(x, ...)

\S4method{summary}{ceData}(object, ...)

\S4method{dim}{ceData}(x)

\S4method{rbind2}{ceData,ceData}(x, y)

\S4method{subset}{ceData}(x, subset, ...)
}
\arguments{
\item{object, x, y, ceData}{\link{ceData-class} object}

\item{subset}{an expression}

\item{\\dots}{parameters}
}
\description{
Methods for \link{ceData-class} object
}
