% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0006_ceDataCons-class.R
\docType{class}
\name{ceDataCons-class}
\alias{ceDataCons-class}
\title{ceDataCons-class}
\description{
Consolidated fishing effort object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{ce}}{Dataframe of 10 columns}
}}

