% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0006_ceDataCons-class.R
\docType{methods}
\name{ceDataCons}
\alias{ceDataCons}
\alias{ceDataCons,ceDataVal,strIni-method}
\alias{ceDataCons,ceDataVal,missing-method}
\alias{ceDataCons,missing,missing-method}
\title{ceDataCons constructors}
\usage{
ceDataCons(object, objStrat, ...)

\S4method{ceDataCons}{ceDataVal,strIni}(object, objStrat, desc = "Unknown stock", ...)

\S4method{ceDataCons}{ceDataVal,missing}(object, desc = "Unknown stock", ...)

\S4method{ceDataCons}{missing,missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object}{\link{ceDataVal-class} object}

\item{objStrat}{stratification \link{strIni-class} object}

\item{desc}{character vector of stock information}
}
\description{
ceDataCons constructors
}
