% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0006_ceDataCons-class.R
\docType{methods}
\name{ce,ceDataCons-method}
\alias{ce,ceDataCons-method}
\alias{head,ceDataCons-method}
\alias{summary,ceDataCons-method}
\alias{dim,ceDataCons-method}
\alias{rbind2,ceDataCons,ceDataCons-method}
\alias{subset,ceDataCons-method}
\title{ceDataCons methods}
\usage{
\S4method{ce}{ceDataCons}(object, ...)

\S4method{head}{ceDataCons}(x, ...)

\S4method{summary}{ceDataCons}(object, ...)

\S4method{dim}{ceDataCons}(x)

\S4method{rbind2}{ceDataCons,ceDataCons}(x, y)

\S4method{subset}{ceDataCons}(x, subset, ...)
}
\arguments{
\item{object, x, y}{\link{ceDataCons-class} object}

\item{subset}{subset expression}

\item{\\dots}{parameters}
}
\description{
ceDataCons methods
}
