% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0005_ceDataVal-class.R
\docType{class}
\name{ceDataVal-class}
\alias{ceDataVal-class}
\alias{ceDataVal}
\alias{ceDataVal,ceData-method}
\alias{ceDataVal,missing-method}
\title{ceDataVal-class}
\usage{
ceDataVal(object, ...)

\S4method{ceDataVal}{ceData}(object, ...)

\S4method{ceDataVal}{missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object, desc, \dots}{parameters}
}
\description{
Fishing effort object
}
\section{Slots}{

\describe{
\item{\code{ce}}{a \link{ceData-class} object}
}}

