% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0007_check-methods.R
\name{checkDataIntegrity}
\alias{checkDataIntegrity}
\alias{checkDataIntegrity,data.frame,data.frame-method}
\title{Check data integrity}
\usage{
checkDataIntegrity(target, current, ...)

\S4method{checkDataIntegrity}{data.frame,data.frame}(target, current, report = FALSE, ...)
}
\arguments{
\item{target, current}{dataframe}

\item{report}{boolean}

\item{\\dots}{parameters}
}
\description{
Check data integrity
}
