% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-methods.R
\docType{methods}
\name{checkNms}
\alias{checkNms}
\alias{checkNms,data.frame,character-method}
\title{Check names in tables}
\usage{
checkNms(object, nms, ...)

\S4method{checkNms}{data.frame,character}(object, nms, ...)
}
\arguments{
\item{dataframe}{}

\item{character}{vectors}
}
\description{
Check names in tables
}
