% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-methods.R
\name{checkpk}
\alias{checkpk}
\title{Check primary keys in tables}
\usage{
checkpk(data, names, type)
}
\arguments{
\item{object}{a \link{ceData-class} object}
}
\value{
a boolean (TRUE if the object is a \link{ceData-class} object, FALSE if not)
}
\description{
Fishing effort data class validity check
}
