% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0007_check-methods.R
\name{checkTys}
\alias{checkTys}
\alias{checkTys,data.frame,list-method}
\title{checkTys}
\usage{
checkTys(object, tys, ...)

\S4method{checkTys}{data.frame,list}(object, tys, ...)
}
\arguments{
\item{object}{data.frame}

\item{tys}{parameters}

\item{data.frame}{parameters}

\item{list}{a list}

\item{\\dots}{parameters}
}
\description{
Methods to check types of columns in tables
}
