% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0008-clData-class.R
\docType{class}
\name{clData-class}
\alias{clData-class}
\title{clData-class}
\description{
Landing object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{cl}}{Dataframe of  columns}
}}

