% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0008-clData-class.R
\docType{methods}
\name{clData}
\alias{clData}
\alias{clData,data.frame-method}
\alias{clData,missing-method}
\title{clData constructors}
\usage{
clData(cl, ...)

\S4method{clData}{data.frame}(cl, desc = "Unknown stock", ...)

\S4method{clData}{missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{cl}{dataframe following the cl slot of a \link{clData-class} object}

\item{desc}{descriptor}

\item{missing}{: create an empty \link{clData-class} object}

\item{\\dots}{parameters}
}
\description{
clData constructors
}
