% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0008-clData-class.R
\docType{methods}
\name{cl}
\alias{cl}
\alias{cl,clData-method}
\alias{head,clData-method}
\alias{summary,clData-method}
\alias{dim,clData-method}
\alias{rbind2,clData,clData-method}
\alias{subset,clData-method}
\title{clData-methods}
\usage{
cl(object, ...)

\S4method{cl}{clData}(object, ...)

\S4method{head}{clData}(x, ...)

\S4method{summary}{clData}(object, ...)

\S4method{dim}{clData}(x)

\S4method{rbind2}{clData,clData}(x, y)

\S4method{subset}{clData}(x, subset, ...)
}
\arguments{
\item{object, x, y, clData}{\link{clData-class} object}

\item{subset}{an expression}

\item{\\dots}{parameters}
}
\description{
Methods for \link{clData-class} object
}
