% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0010_clDataCons-class.R
\docType{class}
\name{clDataCons-class}
\alias{clDataCons-class}
\title{clDataCons-class}
\description{
Consolidated landing object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{cl}}{Dataframe of xx columns}
}}

