% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0010_clDataCons-class.R
\docType{methods}
\name{clDataCons}
\alias{clDataCons}
\alias{clDataCons,clDataVal,strIni-method}
\alias{clDataCons,clDataVal,missing-method}
\alias{clDataCons,missing,missing-method}
\title{clDataCons constructors}
\usage{
clDataCons(object, objStrat, ...)

\S4method{clDataCons}{clDataVal,strIni}(object, objStrat, desc = "Unknown stock", ...)

\S4method{clDataCons}{clDataVal,missing}(object, desc = "Unknown stock", ...)

\S4method{clDataCons}{missing,missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object}{\link{clDataVal-class} object}

\item{objStrat}{stratification \link{strIni-class} object}

\item{desc}{character vector of stock information}

\item{\\dots}{parameters}
}
\description{
clDataCons constructors
}
