% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0010_clDataCons-class.R
\docType{methods}
\name{cl,clDataCons-method}
\alias{cl,clDataCons-method}
\alias{head,clDataCons-method}
\alias{summary,clDataCons-method}
\alias{dim,clDataCons-method}
\alias{rbind2,clDataCons,clDataCons-method}
\alias{subset,clDataCons-method}
\title{clDataCons methods}
\usage{
\S4method{cl}{clDataCons}(object, ...)

\S4method{head}{clDataCons}(x, ...)

\S4method{summary}{clDataCons}(object, ...)

\S4method{dim}{clDataCons}(x)

\S4method{rbind2}{clDataCons,clDataCons}(x, y)

\S4method{subset}{clDataCons}(x, subset, ...)
}
\arguments{
\item{object, x, y}{\link{clDataCons-class} object}

\item{subset}{subset expression}

\item{\\dots}{parameters}
}
\description{
clDataCons methods
}
