% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0009_clDataVal-class.R
\docType{class}
\name{clDataVal-class}
\alias{clDataVal-class}
\alias{clDataVal}
\alias{clDataVal,clData-method}
\alias{clDataVal,missing-method}
\title{clDataVal-class}
\usage{
clDataVal(object, ...)

\S4method{clDataVal}{clData}(object, ...)

\S4method{clDataVal}{missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object}{parameters}

\item{desc}{parameters}

\item{\\dots}{parameters}
}
\description{
Validated landing object
}
\section{Slots}{

\describe{
\item{\code{cl}}{a \link{clData-class} object}
}}

