% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0002_coerceCons.R
\name{coerceCons,data.frame,data.frame-method}
\alias{coerceCons,data.frame,data.frame-method}
\title{rdname coerceCons}
\usage{
\S4method{coerceCons}{data.frame,data.frame}(object, refObject, ...)
}
\description{
rdname coerceCons
}
