% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0002_coerceCons.R
\docType{methods}
\name{coerceCons}
\alias{coerceCons}
\title{coerceCons}
\usage{
coerceCons(object, refObject, ...)
}
\arguments{
\item{object}{dataframe}

\item{refObject}{dataframe}
}
\description{
Convert columns of a data.frame with respect to refObject format
}
