% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0014_utils.R
\name{coerceDataFrameColumns}
\alias{coerceDataFrameColumns}
\alias{coerceDataFrameColumns,data.frame,data.frame-method}
\title{coerceDataFrameColumns}
\usage{
coerceDataFrameColumns(object, refObject, ...)

\S4method{coerceDataFrameColumns}{data.frame,data.frame}(object, refObject, ...)
}
\arguments{
\item{object}{object to coerce}

\item{refObject}{reference object}

\item{\\dots}{parameters}
}
\description{
coerceDataFrameColumns
}
