% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSTlite-package.R
\docType{package}
\name{COSTlite-package}
\alias{COSTlite}
\alias{COSTlite-package}
\title{COSTlite: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Laurent Dubroca \email{laurent.dubroca@gmail.com} (\href{https://orcid.org/0000-0002-1861-0507}{ORCID})

}
\keyword{internal}
