% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\docType{class}
\name{csData-class}
\alias{csData-class}
\title{csData-class}
\description{
Commercial fisheries sampling data object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{tr}}{trip information}

\item{\code{hh}}{haul information}

\item{\code{sl}}{species list information}

\item{\code{hl}}{length frequencies information}

\item{\code{ca}}{age,weight, maturity information}
}}

