% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\docType{methods}
\name{csData}
\alias{csData}
\alias{csData,data.frame,missing,missing,missing,missing-method}
\alias{csData,data.frame,data.frame,missing,missing,missing-method}
\alias{csData,data.frame,data.frame,data.frame,missing,missing-method}
\alias{csData,data.frame,data.frame,data.frame,data.frame,missing-method}
\alias{csData,data.frame,data.frame,data.frame,data.frame,data.frame-method}
\alias{csData,data.frame,missing,missing,missing,data.frame-method}
\alias{csData,missing,missing,missing,missing,missing-method}
\title{csData constructors}
\usage{
csData(tr, hh, sl, hl, ca, ...)

\S4method{csData}{data.frame,missing,missing,missing,missing}(tr, desc = "Unknown stock", ...)

\S4method{csData}{data.frame,data.frame,missing,missing,missing}(tr, hh, desc = "Unknown stock", ...)

\S4method{csData}{data.frame,data.frame,data.frame,missing,missing}(tr, hh, sl, desc = "Unknown stock", ...)

\S4method{csData}{data.frame,data.frame,data.frame,data.frame,missing}(tr, hh, sl, hl, desc = "Unknown stock", ...)

\S4method{csData}{data.frame,data.frame,data.frame,data.frame,data.frame}(tr, hh, sl, hl, ca, desc = "Unknown stock", ...)

\S4method{csData}{data.frame,missing,missing,missing,data.frame}(tr, hh, sl, hl, ca, desc = "Unknown stock", ...)

\S4method{csData}{missing,missing,missing,missing,missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{tr, hh, sl, hl, ca}{dataframe following the slots of a \link{csData-class} object}

\item{desc}{descriptor}

\item{missing}{: create an empty \link{csData-class} object}

\item{\\dots}{parameters}
}
\description{
csData constructors
}
