% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\docType{methods}
\name{tr}
\alias{tr}
\alias{tr,csData-method}
\alias{hh}
\alias{hh,csData-method}
\alias{sl}
\alias{sl,csData-method}
\alias{hl}
\alias{hl,csData-method}
\alias{ca}
\alias{ca,csData-method}
\alias{head,csData-method}
\alias{summary,csData-method}
\alias{dim,csData-method}
\alias{rbind2,csData,csData-method}
\title{csData-methods}
\usage{
tr(object, ...)

\S4method{tr}{csData}(object, ...)

hh(object, ...)

\S4method{hh}{csData}(object, ...)

sl(object, ...)

\S4method{sl}{csData}(object, ...)

hl(object, ...)

\S4method{hl}{csData}(object, ...)

ca(object, ...)

\S4method{ca}{csData}(object, ...)

\S4method{head}{csData}(x, ...)

\S4method{summary}{csData}(object, ...)

\S4method{dim}{csData}(x)

\S4method{rbind2}{csData,csData}(x, y)
}
\arguments{
\item{object, x, y}{\link{csData-class} object}

\item{\\dots}{parameters}
}
\description{
Methods for \link{csData-class} object
}
