% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0013_csDataCons-class.R
\docType{class}
\name{csDataCons-class}
\alias{csDataCons-class}
\title{csDataCons-class}
\description{
Commercial fisheries sampling data consolidated object
}
\section{Slots}{

\describe{
\item{\code{desc}}{Character chain of a descriptor}

\item{\code{tr}}{trip information}

\item{\code{hh}}{haul information}

\item{\code{sl}}{species list information}

\item{\code{hl}}{length frequencies information}

\item{\code{ca}}{age,weight, maturity information}
}}

