% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0013_csDataCons-class.R
\name{csDataCons}
\alias{csDataCons}
\alias{csDataCons,csDataVal,missing-method}
\alias{csDataCons,missing,missing-method}
\title{csDataCons constructors}
\usage{
csDataCons(object, objStrat, ...)

\S4method{csDataCons}{csDataVal,missing}(object, desc = "Unknown stock", ...)

\S4method{csDataCons}{missing,missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object}{\link{csDataVal-class} object}

\item{objStrat}{stratification \link{strIni-class} object}

\item{desc}{character vector of stock information}

\item{\\dots}{parameters}
}
\description{
csDataCons constructors
}
