% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0013_csDataCons-class.R
\name{tr,csDataCons-method}
\alias{tr,csDataCons-method}
\alias{hh,csDataCons-method}
\alias{sl,csDataCons-method}
\alias{hl,csDataCons-method}
\alias{ca,csDataCons-method}
\alias{head,csDataCons-method}
\alias{summary,csDataCons-method}
\alias{dim,csDataCons-method}
\alias{rbind2,csDataCons,csDataCons-method}
\alias{subset,csDataCons-method}
\alias{subsetSpp,csDataCons-method}
\title{csDataCons methods}
\usage{
\S4method{tr}{csDataCons}(object, ...)

\S4method{hh}{csDataCons}(object, ...)

\S4method{sl}{csDataCons}(object, ...)

\S4method{hl}{csDataCons}(object, ...)

\S4method{ca}{csDataCons}(object, ...)

\S4method{head}{csDataCons}(x, ...)

\S4method{summary}{csDataCons}(object, ...)

\S4method{dim}{csDataCons}(x)

\S4method{rbind2}{csDataCons,csDataCons}(x, y)

\S4method{subset}{csDataCons}(x, subset, ...)

\S4method{subsetSpp}{csDataCons}(x, subset, link = TRUE, ...)
}
\arguments{
\item{object, x, y, link}{\link{ceDataCons-class} object}

\item{subset}{subset expression}

\item{\\dots}{parameters}
}
\description{
csDataCons methods
}
