% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0013_csDataCons-class.R
\name{csDataConsCATR}
\alias{csDataConsCATR}
\title{Class constructor for validated objects with only tr & ca tables filled (not to be exported)}
\usage{
csDataConsCATR(object, objStrat, desc = "Unknown stock", ...)
}
\arguments{
\item{object}{a \link{csDataVal-class} object}

\item{objStrat}{stratification \link{strIni-class} object}

\item{desc}{character vector of stock information}

\item{\\dots}{parameter}
}
\description{
Fishing effort data class validity check
}
