% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0012_csDataVal-class.R
\docType{class}
\name{csDataVal-class}
\alias{csDataVal-class}
\alias{csDataVal}
\alias{csDataVal,csData-method}
\alias{csDataVal,missing-method}
\title{csDataVal-class}
\usage{
csDataVal(object, ...)

\S4method{csDataVal}{csData}(object, ...)

\S4method{csDataVal}{missing}(desc = "Unknown stock", ...)
}
\arguments{
\item{object, desc, \dots}{parameters}
}
\description{
Commercial sampling validated object
}
\section{Slots}{

\describe{
\item{\code{cs}}{a \link{csData-class} object}
}}

