% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0000_data.R
\docType{data}
\name{data}
\alias{data}
\alias{sole.cl}
\alias{sole.ce}
\alias{sole.cs}
\alias{data.ca}
\alias{data.sl}
\alias{data.hh}
\alias{data.hl}
\alias{data.tr}
\alias{data.cl}
\alias{data.ce}
\title{Data
Example data}
\format{
An object of class \code{clData} with 931 rows and 22 columns.

An object of class \code{ceData} with 221 rows and 18 columns.

An object of class \code{csData} of dimension c(95, 16) x c(447, 29) x c(293, 17) x c(1409, 16) x c(1627, 31).

An object of class \code{csData} of dimension c(95, 16) x c(447, 29) x c(293, 17) x c(1409, 16) x c(1627, 31).

An object of class \code{data.frame} with 1627 rows and 31 columns.

An object of class \code{data.frame} with 293 rows and 17 columns.

An object of class \code{data.frame} with 447 rows and 29 columns.

An object of class \code{data.frame} with 1409 rows and 16 columns.

An object of class \code{data.frame} with 95 rows and 16 columns.

An object of class \code{data.frame} with 931 rows and 22 columns.

An object of class \code{data.frame} with 221 rows and 18 columns.
}
\usage{
sole.cl

sole.ce

sole.cs

sole.cs

data.ca

data.sl

data.hh

data.hl

data.tr

data.cl

data.ce
}
\description{
Data
Example data
}
\keyword{datasets}
