% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockinfo.R
\name{defstock}
\alias{defstock}
\title{Stock definition}
\usage{
defstock(stock, wg, year, CLr, CSr, path2precatch = "../../precatch.csv")
}
\arguments{
\item{stock, wg}{stock id and ICES WG}

\item{year}{processing year}

\item{CLr}{commercial landing}

\item{CSr}{commercial sampling}
}
\value{
list of 2 objects with stock info
}
\description{
Stock definition and information
}
