% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockinfo.R
\name{findprecatch}
\alias{findprecatch}
\title{Compute precatch}
\usage{
findprecatch(path2precatch, CLr, CSr)
}
\arguments{
\item{path2precatch}{path to precatch file}

\item{CLr}{commercial landing}

\item{CSr}{commercial sampling}
}
\value{
precatch value based on CLr area and spp in CSr
}
\description{
compute precatch from ICES precatch
}
