% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0004_recodevalues.R
\name{mapvalues}
\alias{mapvalues}
\title{Map old values to new values for a vector}
\usage{
mapvalues(x, from, to, warn_missing = FALSE)
}
\arguments{
\item{x}{a vector to be mapped}

\item{from}{a vector of the old values}

\item{to}{a vector of the new values}

\item{warn_missing}{a boolean}
}
\value{
a vector newly mapped
}
\description{
Map old values to new values for a vector
}
