% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0004_recodevalues.R
\name{recFun}
\alias{recFun}
\title{Internal recoding function}
\usage{
recFun(df, field, rec)
}
\arguments{
\item{df}{dataframe}

\item{field}{truc}

\item{rec}{bidule}
}
\value{
df
}
\description{
Internal recoding function
}
