% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataref.R
\docType{data}
\name{reffao}
\alias{reffao}
\title{reffao}
\format{
An object of class \code{data.frame} with 13096 rows and 15 columns.
}
\usage{
reffao
}
\description{
Reference table for species names (2024 version)
}
\keyword{datasets}
