% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0014_utils.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(x, size, replace, ...)
}
\arguments{
\item{x, size, replace}{object}

\item{\\dots}{parameters}
}
\description{
internal function
sample function that returns one value if give data with one value x,
rather than returning sample of vector 1:x , see help file for sample
}
