% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0014_utils.R
\name{slSex}
\alias{slSex}
\title{slSex}
\usage{
slSex(slTab, hlTab)
}
\arguments{
\item{slTab, hlTab}{object}
}
\description{
internal function
method for handle 'sex' conditional key field in SL : an new 'sex' field is inserted in HL, matching SL sex field, and true HL sex field is moved down and renamed as 'lsex'
}
