% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0003_strIni.R
\docType{methods}
\name{strIni-class}
\alias{strIni-class}
\alias{strIni}
\title{Stratification definition object for consolidated objects creation}
\usage{
strIni(
  timeStrata = as.character(NA),
  spaceStrata = as.character(NA),
  techStrata = as.character(NA),
  tpRec = as.list(NA),
  spRec = as.list(NA),
  tcRec = as.list(NA)
)
}
\arguments{
\item{timeStrata}{character vector definition of the temporal stratification}

\item{spaceStrata}{character vector definition of the spatial stratification}

\item{techStrata}{character vector definition of the technical stratification}

\item{tpRec}{a list defining the temporal stratification}

\item{spRec}{a list defining the spatial stratification}

\item{tcRec}{a list defining the technical stratification}
}
\description{
Stratification definition object for consolidated objects creation
}
