% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\name{subsetCOST}
\alias{subsetCOST}
\alias{subset,csData-method}
\title{Subset \link{csData-class} object}
\usage{
subsetCOST(x, subset, ..., table = "tr")

\S4method{subset}{csData}(x, subset, ..., table = "tr", link = FALSE)
}
\arguments{
\item{x}{a \link{csData-class} object}

\item{subset}{subset expression}

\item{table}{the table targeted by the subset}

\item{link}{boolean if TRUE 'table' & ca tables are subset}

\item{\\dots}{parameters}
}
\value{
a \link{csData-class} object, subsetted
}
\description{
subset : one specified table is subset, and other tables are accordingly subset downward and upward (hh, sl , hl subsets don't impact on ca ; ca subset doesn't impact on hh, sl & hl)
}
