% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\name{subsetSpp}
\alias{subsetSpp}
\alias{subsetSpp,csData-method}
\title{subsetSpp}
\usage{
subsetSpp(x, subset, link = FALSE, ...)

\S4method{subsetSpp}{csData}(x, subset, link = FALSE, ...)
}
\arguments{
\item{x, subset, link}{parameters}

\item{\\dots}{parameters}
}
\description{
This method implements subsetting for the raw, the validated and the
consolidated CS objects provided by COSTcore, proceeding specifically on SL
table. This subset only impacts on HL table, and preserves the other tables.
If link, ca is subsetted.
}
