% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockinfo.R
\name{teststock}
\alias{teststock}
\title{Test data availability}
\usage{
teststock(info, CLr, CSr, param)
}
\arguments{
\item{info}{an info object fomr defstock function}

\item{CLr}{commercial landing}

\item{CSr}{commercial sampling}

\item{param}{the parameters}
}
\value{
table of results
}
\description{
Test data availability
}
