% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0006_ceDataCons-class.R
\name{valcecData}
\alias{valcecData}
\title{Validity check function for \link{ceDataCons-class} object}
\usage{
valcecData(object)
}
\arguments{
\item{object}{a \link{ceDataCons-class} object}
}
\value{
a boolean (TRUE if the object is a \link{ceDataCons-class} object, FALSE if not)
}
\description{
Fishing effort data class validity check
}
