% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0001_ceData-class.R
\name{valceData}
\alias{valceData}
\title{Validity check function for \link{ceData-class} object}
\usage{
valceData(object)
}
\arguments{
\item{object}{a \link{ceData-class} object}
}
\value{
a boolean (TRUE if the object is a \link{ceData-class} object, FALSE if not)
}
\description{
Fishing effort data class validity check
}
