% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0010_clDataCons-class.R
\name{valclcData}
\alias{valclcData}
\title{Validity check function for \link{clDataCons-class} object}
\usage{
valclcData(object)
}
\arguments{
\item{object}{a \link{clDataCons-class} object}
}
\value{
a boolean (TRUE if the object is a \link{clDataCons-class} object, FALSE if not)
}
\description{
Landing data class validity check
}
