% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0008-clData-class.R
\name{valclData}
\alias{valclData}
\title{Validity check function for \link{clData-class} object}
\usage{
valclData(object)
}
\arguments{
\item{object}{a \link{clData-class} object}
}
\value{
a boolean (TRUE if the object is a \link{clData-class} object, FALSE if not)
}
\description{
Landing data class validity check
}
