% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0013_csDataCons-class.R
\name{valcscData}
\alias{valcscData}
\title{Validity check function for \link{csDataCons-class} object}
\usage{
valcscData(object)
}
\arguments{
\item{object}{a \link{csDataCons-class} object}
}
\value{
a boolean (TRUE if the object is a \link{csDataCons-class} object, FALSE if not)
}
\description{
Commercial sampling data class validity check
}
