% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0011_csData-class.R
\name{valcsData}
\alias{valcsData}
\title{Validity check function for \link{csData-class} object}
\usage{
valcsData(object)
}
\arguments{
\item{object}{a \link{csData-class} object}
}
\value{
a boolean (TRUE if the object is a \link{csData-class} object, FALSE if not)
}
\description{
Sampling data class validity check
}
