% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataref.R
\docType{data}
\name{wgparam}
\alias{wgparam}
\title{wgparam}
\format{
An object of class \code{data.frame} with 316 rows and 10 columns.
}
\usage{
wgparam
}
\description{
Stock Reference table from ICES (2024 version)
}
\keyword{datasets}
