% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0007_check-methods.R
\name{checkNms}
\alias{checkNms}
\alias{checkNms,data.frame,character-method}
\title{Check names in tables}
\usage{
checkNms(object, nms, ...)

\S4method{checkNms}{data.frame,character}(object, nms, ...)
}
\arguments{
\item{object, }{dataframe,character parameters}

\item{nms}{parameters}

\item{\\dots}{parameters}
}
\description{
Check names in tables
}
