% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0007_check-methods.R
\name{checkpk}
\alias{checkpk}
\title{Check primary keys in tables}
\usage{
checkpk(object, type)
}
\arguments{
\item{object}{a data.frame}

\item{type}{the object type}
}
\value{
a boolean (TRUE if the object is a \link{ceData-class} object, FALSE if not)
}
\description{
Fishing effort data class validity check
}
