% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0002_coerceCons.R
\name{coerceCons}
\alias{coerceCons}
\alias{coerceCons,data.frame,data.frame-method}
\title{coerceCons}
\usage{
coerceCons(object, refObject, ...)

\S4method{coerceCons}{data.frame,data.frame}(object, refObject, ...)
}
\arguments{
\item{object}{dataframe}

\item{refObject}{dataframe}

\item{\\dots}{parameters}

\item{data.frame}{parameters}
}
\description{
Convert columns of a data.frame with respect to refObject format
}
